// packages/frontend/src/typechain-types/contracts/launchpad/Launchpad.ts
export type ProjectDetailsStruct = {
  name: string;
  symbol: string;
  description: string;
  iconUrl: string;
};

export type ProjectDetailsStructOutput = [
  name: string,
  symbol: string,
  description: string,
  iconUrl: string
] & { name: string; symbol: string; description: string; iconUrl: string };

export type LaunchStruct = {
  id: BigNumberish;
  purchaseToken: AddressLike;
  purchaseNftAddress: AddressLike;
  targetRaise: BigNumberish;
  raised: BigNumberish;
  tokenAddress: AddressLike;
  tokenDecimals: BigNumberish;
  tokensToBeEmitted: BigNumberish;
  capPerUser: BigNumberish;
  saleStart: BigNumberish;
  saleEnd: BigNumberish;
  totalUsers: BigNumberish;
  purchaseFormula: AddressLike;
  reserveRatio: BigNumberish;
  claimEnabled: boolean;
  launchpadFee: BigNumberish;
  launchpadFeeToken: BigNumberish;
  details: ProjectDetailsStruct;
};